updated curb intrumentation
